feat(engine)!: factorize prompt logic and make it more reliable
By more reliable, we mean abandon the false promise to have mksls'
terminal output be solely composed of feedback lines in the form:
    (<action>) <link> -> <target>
Indeed, after more research, it seems impossible to do in the general
case, because we need to overwrite prompt lines, which number depends
on the font size, terminal size, user input, ..., all of which can
vary during the program's execution, making the number of lines pretty
much impossible to track in all cases.
The new approach let's prompt lines as they are.